requirements.txt
asgiref==3.8.1
biopython==1.85
Django==5.0.6
django-cors-headers==4.8.0
django-extensions==3.2.3
joblib==1.5.2
numpy==2.3.2
pandas==2.3.2
pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
sqlparse==0.5.0
threadpoolctl==3.6.0
tzdata==2025.2
